Apollo Error Handling Workaround
If there are any errors in a graphql response, Apollo will drop the partial
data and only return the errors. This is a known issue in the Apollo repo
(issue #51). In the meantime, all errors from the ApolloClient have been
moved into the 'data' field under 'errors' so components can access them AND
the partial data. This should be removed when the Apollo issue is closed.